Ejercicio 2

+++ ejercicio1.py
@@ -1,7 +1,7 @@
 sum=0
 n=5
 for i in range(n):
-    current_input = int(input(f'Ingrese voltaje {i+1}: '))
+    current_input = float(input(f'Ingrese voltaje {i+1}: '))
     sum += current_input
 avg = sum/n
 if avg>220:
